+++ contracts/SintropInterface.sol
@@ -2,5 +2,8 @@
 pragma solidity >=0.7.0 <=0.9.0;
 
 interface SintropInterface {
-    function getProducerApprove(address delegate) external view returns(uint);
+    function getProducerApprove(address delegate)
+        external
+        view
+        returns (uint256);
 }
